PJ: fix offset to advance for int types in var args.
Fixes crashes on 64-bit machines when substituting parameters into
dynamic strings, e.g. save flic menu.

Eventually this va_args string formatting business should be replaced
by something simpler and more portable.

+++ src/PJ/util/ftextf.c
@@ -37,7 +37,7 @@ SHORT offset;
 		switch(FT_TYPE(c))
 		{
 			case FT_INT:
-				offset += ssizeof(int);
+				offset += ssizeof(intptr_t);
 				if(c == FT_STARTYPE)  /* '*' star type int args */
 					continue;
 				break;
